changed page num to int

+++ src/main/java/edu/uci/ics/cs221/index/inverted/Location.java
@@ -1,11 +1,11 @@
 package edu.uci.ics.cs221.index.inverted;
 
 public class Location {
-    public short Page;
+    public int Page;
     public short Offset;
 
     Location(int page, int offset){
-        Page = (short) page;
+        Page = page;
         Offset = (short) offset;
     }
